# vectorize preprocessing